baixar_vacances_notoi moved to retriever module
